requirements.txt
quart==0.18.4
werkzeug==2.0.1
quart-cors==0.7.0
aiohttp==3.8.5
hypercorn==0.14.4
uvloop==0.17.0
uvicorn==0.23.2
gunicorn==20.1.0
pillow==9.5.0
python-dotenv==1.0.0